--- a/eth-providers/src/__tests__/e2e/evm-rpc-provider.test.ts
+++ b/eth-providers/src/__tests__/e2e/evm-rpc-provider.test.ts
@@ -73,7 +73,8 @@ describe('getReceiptAtBlock', () => {
   })
 });
 
-describe('all cache', () => {
+// TODO: maybe setup a subway to test
+describe.skip('all cache', () => {
   const provider = EvmRpcProvider.from(ACALA_NODE_URL);
 
   beforeAll(async () => await provider.isReady());