--- a/eth-providers/src/utils/BlockCache.ts
+++ b/eth-providers/src/utils/BlockCache.ts
@@ -1,3 +1,5 @@
+import { Header, SignedBlock } from '@polkadot/types/interfaces';
+import LRUCache from 'lru-cache';
 import { FullReceipt } from './transactionReceiptHelper';
 
 export type TxHashToReceipt = Record<string, FullReceipt>;
@@ -11,6 +13,8 @@ export interface CacheInspect {
 }
 
 export class BlockCache {
+  headerCache: LRUCache<string, Header>;
+  blockCache: LRUCache<string, SignedBlock>;
   blockHashToReceipts: BlockHashToReceipts;
   txHashToReceipt: TxHashToReceipt;
   cachedBlockHashes: string[];
@@ -21,6 +25,8 @@ export class BlockCache {
     this.blockHashToReceipts = {};
     this.cachedBlockHashes = [];
     this.maxCachedBlocks = maxCachedBlocks;
+    this.headerCache = new LRUCache({ max: maxCachedBlocks });
+    this.blockCache = new LRUCache({ max: maxCachedBlocks });
   }
 
   // automatically preserve a sliding window of ${maxCachedBlocks} blocks
@@ -48,6 +54,22 @@ export class BlockCache {
 
   getReceiptAtBlock = (txHash: string, blockHash: string): FullReceipt | null => {
     return this.getAllReceiptsAtBlock(blockHash).find(r => r.transactionHash === txHash) ?? null;
+  }
+
+  setHeaderCache = (blockHash: string, header: Header): void => {
+    this.headerCache.set(blockHash, header);
+  }
+
+  setBlockCache = (blockHash: string, block: SignedBlock): void => {
+    this.blockCache.set(blockHash, block);
+  }
+
+  getHeader = (blockHash: string): Header | undefined => {
+    return this.headerCache.get(blockHash);
+  }
+
+  getBlock = (blockHash: string): SignedBlock | undefined => {
+    return this.blockCache.get(blockHash);
   }
 
   inspect = (): CacheInspect => ({