… eth-rpc-adapter/src/__tests__/e2e/consts.ts