## Context
one single async call usually adds 100-500ms in response time, which is already slower than **ALL** sync operations. So in order to achive best performance, we should set the target to "never trigger any redundant async call", and there are many places that we should cache.

#### already implemented
- [x] storage cache
  - [x] `api.query.*`
- [x] receipts cache
- [x] finalized blockhashes
- [x] current finalized block number/hash 

#### not yet
- [ ] current block number/hash 
- [ ] rpc result cache
  - [ ] `api.rpc.getBlock`
  - [ ] `api.rpc.getHeader`

Although we do have a cache size limit, in practice most ETH RPC calls target most recent couple blocks (deploy to current block, get current block info, check tx receipt at curernt block ...), which means even only 0.1% of latest blocks are cached, 90% of requests can be served.

As a result, after all these caches are implemented, we should be confident that 90%+ of external ETH RPC calls trigger **ZERO** async calls, and are responded almost instantly! This will also greatly reduce node burden at high traffic time by 100x.

## Change
- Added all caches for the above not implemented ones, now block cache has:
  - precomputed receipts cache
  - header cache
  - blockdata cache
- moved `getTransactionRequest` pure function out of base-provider
- fixed block gaslimit to `29990102`
- fixed a small possible timing issue for `isReady` with local mode. Basically make sure `this.(latest)block(number|hash)` are all ready before `isReady` returns, this might add 100ms for ready time, but that's fine since it only happens once at start.

## Test
Added performance tests for `getHeader` and `getblock` to make sure cache is effective. 

`getBlock` is the most complex method that involves all sorts of cache. Previously `getBlock` is only about 3x faster (1000ms => 300ms) for the second query since it's only partically cached. Now it's fully cached, 1000x faster! (1000ms => 1ms) 

![Screenshot 2023-03-01 at 16 46 48](https://user-images.githubusercontent.com/16548786/222090654-6300ccce-2203-4923-9862-e312807a2c63.png)


## Todo
monitor memory, adjust cache size accordingly if needed