@xlc 2023-03-01
We could have a subway node hosted next to eth-rpc-adapter to cache Substrate RPC responses.
We could also have another subway node hosted next to user to cache Eth RPC responses.

So while it is useful to get the basic cache done, but no need to make it exhaustive.