@shunjizhan 2023-03-03
@xlc I removed rpc caches for getBlock and getHeader, and we will handle rpc caches with subway. I think it's still worth tracking latest block number and hash though, it requires no space and is this cacheable (since it's dynamically updating)